Inital working scripts, a new movie and cleanup
Basic changes have been made to helloworld.py and slowmovie.py so that
they are now working with the Inky Impression. Initally the work for
this project was done in a cloned version of the repo so all the working
code has been copied over to here. Also a new test movie has been added to
better show off the colors of the screen and the installer for the
Waveshare screen used by Tom has been removed.