Formalize changes and remove code
The changes that were made to the scripts have been formalized, some
comments have been changed and some unnecessary code has been removed. A
saturation flag has been added to slowmovie.py to set the saturation
level when the movie starts. The default value of 0.25 has been picked
for now but that may change in the future.